test: 테스트 구조 개편 및 도메인 단위·통합 테스트 추가
- 테스트 패키지를 unit(domain/application)과 integration으로 분리
  - unit/domain: 순수 도메인 비즈니스 로직 테스트 (Spring/Mock 없음)
  - unit/application: UseCase 흐름 테스트 (Mock 기반)
  - integration: 실제 DB 테스트 (Testcontainers)
- 각 서비스 도메인 단위 테스트 추가
  - CouponDomainTest: isUsableNow, CouponCache.isValid, CouponIssue.isUsable
  - MemberDomainTest: create() 시 Role.USER 자동 설정
  - PaymentDomainTest: create() 필드 저장 검증
  - StoreDomainTest: Store OPEN, Menu AVAILABLE 자동 설정
- store-service 통합 테스트 추가 (findStoreBasicInfo 커스텀 JPQL 포함)
- 통합 테스트에 flush/clear 추가로 JPA 1차 캐시 우회 실DB 검증
- store-service build.gradle에 Testcontainers 의존성 추가
